requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.2
gunicorn==22.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.0.2
wsproto==1.2.0